% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_externals.R
\name{scale_external_od}
\alias{scale_external_od}
\title{Scale an external trips matrix}
\usage{
scale_external_od(df, counts)
}
\arguments{
\item{df}{Matrix in \code{data_frame} format, with columns \code{origin},
\code{destination}, and flow columns.}

\item{counts}{A \code{data_frame} with a \code{station} column indicating
which external station the count represents. There should be a count for
every flow field in \code{df}.}
}
\value{
A \code{data_frame} of the same structure as \code{df}, with columns
  added representing the scaled OD flows.
}
\description{
In modern big data applications, there is a high probability that the source
OD matrix (potentially including IE trips) will not match external station
counts. For example, StreetLight provides OD matrices that are normalized for
data disclosure and seasonality, and that therefore do not represent the actual
flows in the system. This function takes an input matrix and rescales it to
match counts at an external via an Iterative Proportional Fitting (IPF,
FRATAR, or raking) algorithm.
}

