% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_counts}
\alias{test_counts}
\title{Demonstration external station counts}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.}
\usage{
test_counts
}
\description{
An object of class \code{data_frame} containing 2 observations of 2 variables:
\describe{
  \item{station}{A station ID corresponding to the external stations in
  \code{\link{test_odmatrix}} and \code{\link{test_oddf}}}
  \item{aawdt}{Observed traffic volumes at the external station.}
}
}
\keyword{datasets}

