% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_odmatrix}
\alias{test_odmatrix}
\title{Demonstration OD matrix}
\format{An object of class \code{matrix} with 3 rows and 3 columns.}
\usage{
test_odmatrix
}
\description{
An object of class \code{matrix} containing demonstration OD-flows between
three zones.
}
\keyword{datasets}

