% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_scaling.R
\name{scale_matrix_naive}
\alias{scale_matrix_naive}
\title{Naively rescale a matrix}
\usage{
scale_matrix_naive(A, x, matrix_column_name = NULL,
  scale_column_name = NULL)
}
\arguments{
\item{A}{A matrix in either \code{data_frame} or matrix form.}

\item{x}{If a matrix or a \code{data_frame}, then A will be multiplied by the ratio of
\equation{\sum_{\all i j} x_{ij}} to \equation{\sum_{\all i j} A_{ij}}, such
that \equation{\sum_{\all i j} B_{ij} = \sum_{\all i j} x_{ij}}. If \code{x}}

\item{matrix_column_name}{If \code{A} is a \code{data_frame}, the name of the
column that contains the values requiring scaling.}

\item{scale_column_name}{If \code{x} is a \code{data_frame}, the name of the
column that contains the values for scaling.}
}
\value{
Either a matrix or a numeric vector (determined by type of
  \code{A}), with scaled values.
}
\description{
Sometimes it may be necessary to scale a matrix, as in \equation{B = xA} where
}

