% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Untitled.R, R/scale_matrix_naive.R
\name{scale_matrix_naive}
\alias{scale_matrix_naive}
\title{Naive Scaling of a Matrix}
\usage{
scale_matrix_naive(A, scale, matrix_field = NULL)

scale_matrix_naive(A, scale, matrix_field = NULL)
}
\arguments{
\item{A}{A matrix in either \code{data_frame} or \code{matrix} format.}

\item{scale}{A constant that the new matrix will be scaled to.}

\item{A}{A matrix in either \code{data.frame} or \code{matrix} format.}

\item{scale}{A constant that the new matrix will be scaled to.}
}
\value{
An object of \code{class(A)} with rescaled values.

An object of \code{class(A)} with rescaled values.
}
\description{
This function adjusts the scale \eqn{\sum_{ij}m_{ij}} of a matrix to be
a scalar value by multiplying all values of a matrix by a uniform value.

This function adjusts the scale \eqn{\sum_{ij}m_{ij}} of a matrix to be
a scalar value by multiplying all values of a matrix by a uniform value.
}
\examples{
scale_matrix_naive(test_odmatrix, 1500)
scale_matrix_naive(test_oddf, 25125, "volume")
}

